feat: go to insights can also submit mood
