+++ src/app/moods/components/MoodsDisplay.tsx
@@ -2,11 +2,13 @@
 
 import Cube from "./Cube";
 import SliderBox from "./SliderBox";
-import MoodButtons from "./MoodButtons";
+// import MoodButtons from "./MoodButtons";
 import { useDatabase } from "@/context/DatabaseContext";
 import { useState } from "react";
 import { Datum } from "plotly.js";
 import Modal from "@/ui/shared/Modal";
+import { useRouter } from "next/navigation";
+import Button from "@/ui/shared/Button";
 
 export interface NeurochemState {
   dopamine: Datum;
@@ -15,8 +17,10 @@ export interface NeurochemState {
 }
 
 export default function MoodsDisplay() {
+  const router = useRouter();
   const database = useDatabase();
   const [modalOpen, setModalOpen] = useState(false);
+  const [insightsModalOpen, setInsightsModalOpen] = useState(false);
   const [neuroState, setNeuroState] = useState<NeurochemState>({
     dopamine: 1,
     serotonin: 1,
@@ -87,16 +91,50 @@ export default function MoodsDisplay() {
     label: "ok",
     action: () => setModalOpen(false),
   };
+  const insightsForwardButton = {
+    label: "Yes",
+    action: () => {
+      setInsightsModalOpen(false);
+      submitMood();
+      router.push("/insights");
+    },
+  };
+  const insightsBackButton = {
+    label: "No",
+    action: () => {
+      setInsightsModalOpen(false);
+      router.push("/insights");
+    },
+  };
 
   return (
     <>
       <Cube neuroState={neuroState} />
       <SliderBox handleChange={handleChange} neuroState={neuroState} />
-      <MoodButtons submitMood={submitMood} />
+
+      <div className="flex justify-between w-10/12 max-w-xl m-auto">
+        <Button
+          label="Save"
+          className="mt-2 px-3 py-1 bg-twd-primary-purple text-white rounded"
+          onClick={() => submitMood()}
+        />
+        <Button
+          label="Go to Insights"
+          className="mt-2 px-3 py-1 bg-gray-700 text-white rounded"
+          onClick={() => setInsightsModalOpen(true)}
+        />
+      </div>
+
       <Modal
         modalOpen={modalOpen}
         forwardButton={forwardButton}
         title={"You've submitted your mood!"}
+      />
+      <Modal
+        modalOpen={insightsModalOpen}
+        forwardButton={insightsForwardButton}
+        backButton={insightsBackButton}
+        title={"Save mood before continuing?"}
       />
     </>
   );
